Move wait page redirect to authenticator
